dashboard_components/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
ansi2html==1.8.0
async-timeout==4.0.3
attrs==23.1.0
bidict==0.22.1
botocore==1.31.50
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cryptography==41.0.3
dash==2.13.0
dash-bootstrap-components==1.4.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.2.5
frozenlist==1.4.0
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
multidict==6.0.4
nest-asyncio==1.5.7
numpy==1.25.2
packaging==23.1
paho-mqtt==1.6.1
pandas==2.1.0
plotly==5.16.1
promise==2.3
pycparser==2.21
PyMySQL==1.1.0
python-dateutil==2.8.2
python-decouple==3.8
python-engineio==4.7.0
python-socketio==5.7.2
pytz==2023.3.post1
requests==2.28.2
retrying==1.3.4
s3transfer==0.6.2
six==1.16.0
tenacity==8.2.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
websockets==9.1
Werkzeug==2.2.3
yarl==1.9.2
